+++ client/package.json
@@ -92,7 +92,9 @@
     "typescript": "5.7.2",
     "uuid": "^11.0.3",
     "zod": "^3.24.1",
-    "zustand": "^5.0.2"
+    "zustand": "^5.0.2",
+    "@google/generative-ai": "^0.24.1",
+    "@smastrom/react-rating": "^1.5.0"
   },
   "devDependencies": {
     "@faker-js/faker": "^9.3.0",
