+++ client/src/app/evaluator/form-builder/dashboard/form/builder/[formId]/page.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import FormBuilder from "../../../_components/FormBuilder";
-import BuilderContextProvider from "context/builder-provider";
+import BuilderContextProvider from "@/context/builder-provider";
 
 const Builder = () => {
   return (
