+++ client/src/components/form-builder/blocks/HeadingBlock.tsx
@@ -25,7 +25,7 @@ import {
   FormCategoryType,
   ObjectBlockType,
 } from "@/@types/form-block.type";
-import { fontSizeClass, fontWeightClass } from "@/constant";
+import { fontSizeClass, fontWeightClass } from "@/constants/form-builder";
 import { Input } from "../ui/input";
 import { useBuilder } from "@/context/builder-provider";
