+++ client/src/components/form-builder/blocks/ParagraphBlock.tsx
@@ -23,7 +23,7 @@ import {
   FormCategoryType,
   ObjectBlockType,
 } from "@/@types/form-block.type";
-import { fontSizeClass, fontWeightClass } from "@/constant";
+import { fontSizeClass, fontWeightClass } from "@/constants/form-builder";
 import { z } from "zod";
 import { useBuilder } from "@/context/builder-provider";
 import { Textarea } from "../ui/textarea";
